+++ http4k-realtime-core/src/main/kotlin/org/http4k/filter/coreExtensions.kt
@@ -26,3 +26,9 @@ fun ServerFilters.InitialiseWsRequestContext(contexts: Store<RequestContext>) = 
         }
     }
 }
+
+fun ServerFilters.SetWsSubProtocol(subprotocol: String) = WsFilter { next ->
+    {
+        next(it).copy(subprotocol = subprotocol)
+    }
+}