+++ http4k-realtime-core/src/test/kotlin/org/http4k/filter/coreExtensionsTest.kt
@@ -59,4 +59,11 @@ class CoreExtensionsTest {
 
         assertThat(found.get(), equalTo(credentials))
     }
+
+    @Test
+    fun `can set subprotocol on WsResponse`() {
+        val handler = ServerFilters.SetWsSubProtocol("foobar")
+            .then { WsResponse { _ -> } }
+        assertThat(handler(Request(GET, "/")).subprotocol, equalTo("foobar"))
+    }
 }