+++ http4k-server/jetty/src/main/kotlin/org/http4k/server/jettyInfra.kt
@@ -43,8 +43,9 @@ fun WsHandler.toJettyWsHandler() = WebSocketUpgradeHandler(WebSocketComponents()
 
 fun WsHandler.toJettyNegotiator() = object : WebSocketNegotiator.AbstractNegotiator() {
     override fun negotiate(negotiation: WebSocketNegotiation): FrameHandler {
-        val request = negotiation.request.asHttp4kRequest()!!
-        return Http4kWebSocketFrameHandler(this@toJettyNegotiator(request), request)
+        val consumer = this@toJettyNegotiator(negotiation.request.asHttp4kRequest()!!)
+        consumer.subprotocol?.also { negotiation.subprotocol = it }
+        return Http4kWebSocketFrameHandler(consumer)
     }
 }